messing with cmake continues
Finally seeing light at the end of the tunnel. *.ui file needs to be
wrapped first before adding it as an executable. Wrapping macro puts it
into build directory, so current binary directory must be included for
ui_*h file to be found.

qt5_use_modules works only with newer version of cmake. Still working on
getting it to work with cmake under 2.8.8

Building under Windows is a whole different story.